added role system, modify migrations and all endpoints

+++ migrations/20241127151311_user_add_avatar.down.sql
@@ -1,1 +1,6 @@
-ALTER TABLE users DROP COLUMN IF EXISTS avatar;
+DO $$
+BEGIN
+    IF EXISTS (SELECT 1 FROM information_schema.tables WHERE table_name = 'users') THEN
+        ALTER TABLE users DROP COLUMN IF EXISTS avatar;
+    END IF;
+END $$;
